[refactor] 5단계
