+++ src/main/java/racingcar/domain/RacingCar.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import racingcar.pattern.NumberGenerator;
