+++ src/test/java/racingcar/CarTest.java
@@ -4,6 +4,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import racingcar.domain.Car;
+import racingcar.domain.CarName;
+import racingcar.domain.Position;
 
 import static org.assertj.core.api.Assertions.assertThat;
