+++ src/test/java/racingcar/PositionTest.java
@@ -2,6 +2,7 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Position;
 
 import static org.assertj.core.api.Assertions.assertThat;
